Relay double-spends, subject to anti-DOS
As an alert, relay at most one transaction that respends a prevout already
spent in an unconfirmed transaction in this node's mempool.

As before, respends are not added to the mempool.

Anti-Denial-of-Service-Attack provisions:
 - Use a bloom filter to relay only one respend per mempool prevout
 - Rate-limit respend relays to a default of 100 thousand bytes/minute
 - Define tx2.IsEquivalentTo(tx1): equality when scriptSigs are not considered
 - Do not relay these equivalent transactions
 - Do not allow relayed respends into our wallet unless they conflict with it,
   even if they pay us.  This addresses a resource exhaustion attack

CWallet::SyncMetaData is changed to sync only to equivalent transactions.

Remove an unused variable declaration in txmempool.cpp.