Anti-DoS / IP Prio: generify the code a bit and change Tor URL
Feedback from a Tor developer gave a better URL to use. This change also means Tor IPs won't be fetched at all if you've got an unrouteable IP. The code now expects a simple list of IPs with comment lines starting with a hash. Comments may not be on the same line as an IP.