Limit number of transactions in the memory pool
Implement a simple strategy to keep the memory pool from consuming
all of memory, by limiting the number of tranasctions stored in
the memory pool.

By default, the limit is 10-blocks-worth of average-sized (500-byte)
transactionsl (20,000 transactions for 1MB blocks). Command-line
option -maxmempooltx overrides the default (and is useful for testing).

When the limit is hit, low-priority transactions are dropped
immediately. Higher priority transactions are added, and
a random transaction (and their descendants) are evicted
from the mempool.

BUT, transactions generated or submitted by ourselves (via
sendtoaddress/sendmany/sendrawtransation) are always allowed into
the mempool and are never evicted. And transactions marked as
important via the prioritsetransaction RPC call are also never
evicted.