Cleanup the NetworkStyle a little.
The class was created using a scopedPointer so it would be deleted
when it went out of scope at the end of the method.
A much easier way to do this is to use an on-stack instance and that
also means we don't pass a pointer to the constructor of various classes
having unclear ownership and lifetime.

I also removed many usages of pixmap which is very expensive to use
for image manipulations.  A pixmap implicitly uploads its data to the
graphics card memory, so only use it for diplaying purposes!

Last I ran the bitcoin.png through an optimizer which reduced the size
by about 30%