Remove low feerate txes when full
This change implements a mempool/relay policy that limits the
footprint of the memory pool.

After each transaction added to mempool, TrimToSize is called to
remove the lowest-feerate transactions, until dynamic memory
usage is below the limit set by -maxmempool.

The effective feerate for a transaction is defined as

fee(transaction) / (size(transaction) + size(ancestors))

This change copies some code directly from bitcoin/bitcoin
794a8cec5db84fde1cce82ada51740070ec188ac (Matt Corallo) but
omits the dynamic minimum fee part of that change, in addition to
differences in policy.