Rewrite CreateNewBlock
Use the score index on the mempool to only add sorted txs in order.
Remove much of the validation while building the block, relying on
mempool to be consistent and only contain txs that can be mined.

The mempool is assumed to be consistent as far as not containing
txs which spend non-existent outputs or double spends, and scripts
are valid.  Finality of txs is still checked (except not coinbase
maturity, assumed in mempool). Still TestBlockValidity in case
mempool consistency breaks and return error state if an invalid
block was created.

Unit tests are modified to realize that invalid blocks can now
be constructed if the mempool breaks its consistency assumptions
and also updated to have the right fees, since the cached value
is now used for block construction.

Additional changes for XT:
- The score index used is XT's child-pays-for-all-ancestors index
- The priority section is removed
- Support for the -blockminsize argument is removed