Corrections to bad-chain alert triggering
Three corrections are made to the bad-chain alert triggering mechanism:
the sample interval is reduced to 130 minutes (from 240 minutes),
oversampling is eliminated, and the test is changed to make correct use
of the poisson CDF.

A sample interval of 130 minutes is chosen as the smallest interval
capable of detecting a too-few-blocks event with a false positive rate
of once in 50 years.

It is critical that a sample be taken only every 130 minutes, otherwise
the test will catch events that span sample intervals, which it is
designed to miss.  Oversampling is eliminated by changing the
PartitionCheck funtion to return the interval at which it must be run,
and scheduling it at that interval.

A much smaller bias in the other direction is also eliminated.  The
poisson PDF was being used to check the likelihood of a specific
observation, a method which is incorrectly sensitive to the size of the
sample inteval.  This is corrected by using the poisson CDF and doing a
separate statistical check for the high and low cases.

The effect of the third correction is small because, in practice, the
targeted PDF levels accounted for almost all of the probability mass
in the neighborhood of the test.