Extend checkpointing to a safe distance in the past
Checkpoints are hardcoded points that are shipped in the client and we
skip checking of script validity if they come before a checkpoint.

This change builds on that and extends the no-check requirement from
known good points to acceptable points.

It would be absurd to have more than 100 block reorg, the cost of such
an event would be catastrophic in itself (it is the reason why we have
that coinbase spending limit there), but lets be certain and make it
one day (approx 144 blocks) at minimum.
Additionally, to avoid predictability, make a node have the amount of
script-checking-threads as a multiplication factor.

Conflicts:
	src/main.cpp