SPV: Push unconfirmed ancestors to peer. SPV test.
Peer sets connection bloom filter nFlags bit
BLOOM_ANCESTOR_UPDATE_BIT = 4 to request its filter be
automatically updated with the txids of unconfirmed ancestors of
transactions that are already contained in the filter.

The filter is updated with unconfirmed ancestors in the two places
that filtered results for unconfirmed transactions are provided:
transaction relay, and the "mempool" p2p response.  The results
also include the ancestors added.

The included test spv.py relies on new bloom filter support
added to mininode.py.  Additional tests of SPV server
functionality can be added in the future.