De-prioritize IPs with 2 or more connections.
The goal is to force an attacker to spread out with multiple IPs
to exaust connection slots, making the attack more expensive.

This is done by creating a single-ip-ipgroup for every node that does
not belong to any other ipgroups. For every new connection from the same
ip, that group gets lower priority. When a node disconnects, the
priority increases again.