Direct header announcements (BIP130)
This work is based on @sdaftuar work in Bitcoin Core PR #6494 (9fa8c48).
This work is based on @dgenr8 cherry picks in Bitcoin XT #137

This implements BIP130. This implementation is a refactor/rewrite on
previous work done in Bitcoin Core.

This work supports headersfirst together with thin blocks. It is refactored to support unittests,
and unittests have been written for most of the logic.