Always request headers on block announcement
When blocks to announce exceed MAX_BLOCKS_TO_ANNOUNCE, nodes revert to
announcing only the chain tip with a single inv.

Direct fetching on this inv will lead to unconnected headers. While
occasional unconnected headers are fine, if the out-of-order inv
announcement is followed by many others, such as when instamining on
regtest, this will cause us to misbehave the sending node.

To mitigate this, always request headers first. This will not affect
performance in normal case, because blocks are useually received as compact
block announcements.