Impose a minimum transaction size of 100 bytes after the Nov, 15 2018 HF
Summary:
This ensure we are not vulnerable to the leaf node weakness in bitcoin's merkle tree design anymore and give us room to evolve the protocol in the future (or not).

For more information about the weakness, please read https://bitslog.wordpress.com/2018/06/09/leaf-node-weakness-in-bitcoin-merkle-tree-design/ ( http://archive.is/xA9mW ).

Test Plan:
Added unit test that ensure transaction size is properly checked after activation.
Added an integration test to check for activation.

Reviewers: #bitcoin_abc, Mengerian, jasonbcox

Reviewed By: #bitcoin_abc, Mengerian, jasonbcox

Subscribers: jasonbcox, Mengerian, teamcity

Differential Revision: https://reviews.bitcoinabc.org/D1611