Introduce BlockBuilder interface
The interface replaces use of CBlockTemplate. The interface allows for
building different representations of a new block without going through
CBlock.

Before, transactions in 'getblocktemplate' would pass two steps:
    Mempool -> (copy to) CBlock -> (copy to) JSON

Now it is:
    Mempool -> (copy to) JSON

For now, for sanity check, the JSON is converted to CBlock for the
purpose of validations. This should be skipped in the future.

Also fixes a bug where sigop and txfees are counted wrong in
'getblocktemplate' after CTOR.